Refactor: moved CNT to we_control.h. Moved wpeOptionSection to we_opt.c.
