+++ src/we_opt.h
@@ -5,6 +5,7 @@
 
 #include "config.h"
 #include "edit.h"
+#include "we_control.h"
 #include "globals.h"
 #include "we_main.h"
 #include "we_mouse.h"
